+++ CHANGELOG.md
@@ -12,6 +12,7 @@ system. Old cached data will be converted to the new format automatically.
 
 * `disconnect` event removed in favor of `disconnected`
 * `remoteStorage.claimAccess` removed in favor of `remoteStorage.access.claim`
+* `BaseClient#use` and BaseClient#release removed in favor of `BaseClient#cache`
 
 ### New features